feat(voice,people): approve in conversation, and keep shared contacts between projects (#1082)
* feat(cli): pin a Codex-style composer so harness input never mixes into output

The coding harness only owned the TTY between turns, so typing during a
stream echoed into the agent's output. Keep a dedicated field on screen
for the whole session, queue follow-ups while busy, and interrupt with esc.


* fix(voice): tell Henry which device is listening

A phone call still assumed Command Center, so he kept saying the rest was on screen and switching Mac tabs. Keep the auth principal, name the surface, no-op desktop navigation from phone/watch, and copy on the listening device.


* fix(watch): compile VoiceOrbView into the Watch target

Watch chat referenced the orb without target membership, and the WCSession
delegate hopped session/replyHandler onto MainActor, which Swift 6 rejects.


* fix(cli): satisfy clippy on the pinned composer (Default, slices, dead code)


* fix(watch): use a 32-bit hash constant in VoiceOrbView

watchOS Int cannot hold 2654435761, so the orb failed to compile and Watch
chat could not see the type.


* fix(ios): box WCSession replyHandler so HubWatchRelay compiles under Swift 6


* fix(ios): use ai.permagent.ios so this team can sign device builds

ai.permagent.mobile is registered to another Apple Developer team, so archives and device installs fail until the bundle ID matches Z58XSBM78S.


* fix(watch): flatten the 1024 icon and add Hey Siri, talk to Henry

App Store rejects Watch marketing icons with an alpha channel. Siri now
opens live voice on iPhone (Watch chat on the companion) via an App Shortcut,
with Henry registered as a spoken name for the app.


* fix(watch): use the real Permagent logo for the App Store icon

The Watch 1024 asset was the transparent glyph flattened onto black. Point
it at public/PermagentIcon.png (opaque RGB on the brand field) so App Store
validation and the Dock/home-screen mark match.


* feat(chat,people): approve in the conversation, and keep shared contacts between projects

A parked decision is Jesse's to answer on the channel already in use, so yes/no in chat or voice should not send you to the Inbox. People edits go through the live Brain, and the graph reorients so a person on two projects sits between them.


* fix(watch): bundle the companion again so TestFlight installs it with the phone

WKRunsIndependentlyOfCompanionApp made TestFlight list the Watch app as
installable separately. The Watch still needs the iPhone to reach the hub,
so it is a bundled companion.


* test: refresh Decision Inbox prompt snapshots, and drop the developer name from comments


---------

+++ ios/PermagentMobile/PermagentMobile/VoiceOrbView.swift
@@ -50,7 +50,9 @@ private func makeSphere() -> [Pt] {
         let rad = max(0, 1 - y * y).squareRoot()
         let th = golden * Double(i)
         out.append(Pt(x: cos(th) * rad, y: y, z: sin(th) * rad,
-                      seed: Double((UInt64(i) &* 2_654_435_761) % 1000) / 1000))
+                      // Knuth multiplicative hash as UInt32 — Int is 32-bit on
+                      // watchOS, and this constant does not fit in Int32.
+                      seed: Double((UInt32(i) &* 2_654_435_761) % 1000) / 1000))
     }
     return out
 }
@@ -102,7 +104,7 @@ struct VoiceOrbView: View {
     var speaking: Bool
     /// Thinking has no audio, so the orb swells on its own rather than flatlining.
     var thinking: Bool
-    /// Drawn diameter. Watch uses the file-level default (96); iOS keeps 300.
+    /// Render size. iOS conversation uses 300; Watch chat is ~88.
     var diameter: CGFloat = ORB_SIZE
 
     // Motion integrator state lives in a reference type, NOT @State. Writing
